third_party/WebKit/LayoutTests/xr/xrPresentationContext_preserves_canvas.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script>
let width = 10;
let height = 10;

test(function() {
  let canvas = document.createElement("canvas");
  canvas.width = width;
  canvas.height = height;
  let ctx = canvas.getContext('xrpresent');
  let dstCanvas = ctx.canvas;
  assert_equals(dstCanvas.width, width);
  assert_equals(dstCanvas.height, height);
}, "Test that ctx.canvas on a XRPresentationContext returns the original canvas");

</script>
